Implement Wine wrapper workaround
Fightcade is hard coded to only look in `/usr/bin` and `/usr/local/bin`
for wine. Fortunately for us it also seems to look in
`${DIR}/../Resources/wine.sh` for a wine wrapper script to execute in
place of those.

For our use case this lives at `/app/extra/Resources/wine.sh`, so we
create this file during the `apply_extra` step